[fix] fix the instruction OP_S(L|R)LI

+++ src/cpu/execution/exe_inst_32.rs
@@ -80,11 +80,11 @@ pub fn exe_inst(inst: &Instruction, cpu: &mut CPU) {
         },
         OP_SLLI   => {
             cpu.reg[inst.rd.unwrap() as usize] =
-                cpu.reg[inst.rs1.unwrap() as usize] << inst.imm.unwrap() as i32;
-        },
-        OP_SRLI   => {
-            cpu.reg[inst.rd.unwrap() as usize] =
-                cpu.reg[inst.rs1.unwrap() as usize] >> inst.imm.unwrap() as i32;
+                ((cpu.reg[inst.rs1.unwrap() as usize] as u32) << inst.imm.unwrap()) as i32;
+        },                                                
+        OP_SRLI   => {                                    
+            cpu.reg[inst.rd.unwrap() as usize] =          
+                ((cpu.reg[inst.rs1.unwrap() as usize] as u32) >> inst.imm.unwrap()) as i32;
         },
         OP_ADD    => {
             cpu.reg[inst.rd.unwrap() as usize] =
